… src/newgrf_airporttiles.cpp
… src/newgrf_commons.cpp
… src/newgrf_engine.cpp
… src/newgrf_house.cpp
… src/newgrf_industries.cpp
… src/newgrf_object.cpp
… src/newgrf_roadstop.cpp
… src/newgrf_station.cpp
… src/object_cmd.cpp
… src/object_gui.cpp
… src/picker_gui.h
… src/saveload/engine_sl.cpp
… src/saveload/waypoint_sl.cpp
… src/station_cmd.cpp
… src/table/newgrf_debug_data.h
… src/town_cmd.cpp
… src/town_gui.cpp
… src/vehicle.cpp