<!--
Commit message:

- Please use Feature / Add / Change / Fix for player-facing changes. E.g.: "Feature: My cool new feature".
- Please use Feature / Add / Change / Fix followed by "[NewGRF]" or "[Script]" for moddable changes. E.g.: "Feature: [NewGRF] My cool new NewGRF addition".
- Please use Codechange / Codefix for developer-facing changes. E.g.: "Codefix #1234: Validate against nullptr properly".

See https://github.com/OpenTTD/OpenTTD/blob/master/CODINGSTYLE.md#commit-message for more details.
-->

## Motivation / Problem

EngineOverrideManager records the grfid and localid of each engine, so that after save-load the correct engines are used again.

The looking up the EngineID, the complete list is traversed to find the entry by grfid and localid. As the list gets longer this takes more an more time.

An example of this Iron Horse which has approximately 10,000 engines.

<!--
Describe here shortly
* For bug fixes:
    * What problem does this solve?
    * If there is already an issue, link the issue, otherwise describe the problem here.
* For features or gameplay changes:
    * What was the motivation to develop this feature?
    * Does this address any problem with the gameplay or interface?
    * Which group of players do you think would enjoy this feature?
-->


## Description

Reorganise the list into 4 lists (one for each vehicle type) sorted by grfid and localid. This is built in a pre-sorted manner so that building the list itself can use a binary search instead of a linear scan. The overhead of moving the elements in the vector is tiny compared to performing a linear scan.

The original way of accessing the list with EngineID as the index is no longer possible, and now requires a linear scan to find it. Therefore as a prerequisite, some places that used to use the mapping now use the Engine's grfprops instead, which gains a grfid field.

These changes have no effect when no NewGRFs are loaded as in that case there no lookups during the game.

The performance impact depends on how often NewGRFs access information that needs the lookup, but in my testing there's an improvement of nearly 10,000x. With Iron Horse and several other vehicle NewGRFs loaded:

|       | Master | This PR |
| ----- | ------- | -------- |
| start game with iron horse | 865,156µs | 249µs | 
| per 100,000 calls | 637,596µs | 77µs |
| total over loading and 1 year | 19,765,466µs | 2,379µs |



<!--
Describe here shortly
* For bug fixes:
    * How is the problem solved?
* For features or gameplay changes:
    * What does this feature do?
    * How does it improve/solve the situation described under 'motivation'.
-->


## Limitations

I didn't change the saveload format, so saving requires building a temporary list in EngineID order, and loading requires sorting the list into grfid/localid order (although the latter would always be needed for existing savegames.)

Loading now uses `_engine_mngr.SetID()` directly per entry, so it's automatically inserted into the correct order.

<!--
Describe here
* Is the problem solved in all scenarios?
* Is this feature complete? Are there things that could be added in the future?
* Are there things that are intentionally left out?
* Do you know of a bug or corner case that does not work?
-->


## Checklist for review

Some things are not automated, and forgotten often. This list is a reminder for the reviewers.
* The bug fix is important enough to be backported? (label: 'backport requested')
* This PR touches english.txt or translations? Check the [guidelines](https://github.com/OpenTTD/OpenTTD/blob/master/docs/eints.md)
* This PR affects the save game format? (label 'savegame upgrade')
* This PR affects the GS/AI API? (label 'needs review: Script API')
    * ai_changelog.hpp, game_changelog.hpp need updating.
    * The compatibility wrappers (compat_*.nut) need updating.
* This PR affects the NewGRF API? (label 'needs review: NewGRF')
    * newgrf_debug_data.h may need updating.
    * [PR must be added to API tracker](https://wiki.openttd.org/en/Development/NewGRF/Specification%20Status)